Separate dgva and eva util and metadata packages so unnecessary classes aren't imported and cause problems
