+++ eva-lib/src/main/java/uk/ac/ebi/eva/lib/MultiMongoFactoryConfiguration.java
@@ -25,8 +25,8 @@ import org.springframework.context.annotation.Profile;
 import org.springframework.data.mongodb.MongoDbFactory;
 
 import uk.ac.ebi.eva.lib.configuration.SpringDataMongoDbProperties;
-import uk.ac.ebi.eva.lib.utils.DBAdaptorConnector;
-import uk.ac.ebi.eva.lib.utils.MultiMongoDbFactory;
+import uk.ac.ebi.eva.lib.eva_utils.DBAdaptorConnector;
+import uk.ac.ebi.eva.lib.eva_utils.MultiMongoDbFactory;
 
 import java.io.IOException;
