+++ eva-lib/src/main/java/uk/ac/ebi/eva/lib/eva_utils/DBAdaptorConnector.java
@@ -1,14 +1,11 @@
 /*
- * European Variation Archive (EVA) - Open-access database of all types of genetic
- * variation data from all species
- *
- * Copyright 2014-2017 EMBL - European Bioinformatics Institute
+ * Copyright 2017 EMBL - European Bioinformatics Institute
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *   http://www.apache.org/licenses/LICENSE-2.0
+ *      http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -16,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.ac.ebi.eva.lib.utils;
+package uk.ac.ebi.eva.lib.eva_utils;
 
 import com.mongodb.MongoClient;
 import com.mongodb.MongoClientOptions;
