+++ eva-lib/src/main/java/uk/ac/ebi/eva/lib/eva_utils/MultiMongoDbFactory.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016 EMBL - European Bioinformatics Institute
+ * Copyright 2017 EMBL - European Bioinformatics Institute
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.ac.ebi.eva.lib.utils;
+package uk.ac.ebi.eva.lib.eva_utils;
 
 import com.mongodb.DB;
 import com.mongodb.MongoClient;
