+++ eva-lib/src/main/java/uk/ac/ebi/eva/lib/metadata/eva/StudyEvaproDBAdaptor.java
@@ -1,11 +1,11 @@
 /*
- * Copyright 2014-2017 EMBL - European Bioinformatics Institute
+ * Copyright 2017 EMBL - European Bioinformatics Institute
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *   http://www.apache.org/licenses/LICENSE-2.0
+ *      http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -13,43 +13,44 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.ac.ebi.eva.lib.metadata;
+package uk.ac.ebi.eva.lib.metadata.eva;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Component;
-import uk.ac.ebi.eva.lib.entities.DgvaStudyBrowser;
+
+import uk.ac.ebi.eva.lib.metadata.StudyDBAdaptor;
 import uk.ac.ebi.eva.lib.models.VariantStudy;
+import uk.ac.ebi.eva.lib.entities.EvaStudyBrowser;
 import uk.ac.ebi.eva.lib.utils.QueryOptions;
 import uk.ac.ebi.eva.lib.utils.QueryResult;
-import uk.ac.ebi.eva.lib.repositories.DgvaStudyBrowserRepository;
-import uk.ac.ebi.eva.lib.utils.DgvaDBUtils;
+import uk.ac.ebi.eva.lib.repositories.EvaStudyBrowserRepository;
 
 import java.util.ArrayList;
 import java.util.List;
 
-/**
- * Created by jorizci on 03/10/16.
- */
+import static uk.ac.ebi.eva.lib.eva_utils.EvaproDbUtils.getSpeciesAndTypeFilters;
+
 @Component
-public class StudyDgvaDBAdaptor implements StudyDBAdaptor {
+public class StudyEvaproDBAdaptor implements StudyDBAdaptor {
 
     @Autowired
-    private DgvaStudyBrowserRepository dgvaStudyBrowserRepository;
+    private EvaStudyBrowserRepository evaStudyBrowserRepository;
 
     @Override
     public QueryResult getAllStudies(QueryOptions queryOptions) {
         long start = System.currentTimeMillis();
-        Specification filterSpecification = DgvaDBUtils.getSpeciesAndTypeFilters(queryOptions);
-        List<DgvaStudyBrowser> dgvaStudies = dgvaStudyBrowserRepository.findAll(filterSpecification);
+        Specification filterSpecification = getSpeciesAndTypeFilters(queryOptions);
+        List<EvaStudyBrowser> dgvaStudies = evaStudyBrowserRepository.findAll(filterSpecification);
         List<VariantStudy> variantstudies = new ArrayList<>();
-        for (DgvaStudyBrowser dgvaStudy : dgvaStudies) {
-            variantstudies.add(dgvaStudy.generateVariantStudy());
+        for (EvaStudyBrowser dgvaStudy : dgvaStudies) {
+            if (dgvaStudy != null) {
+                variantstudies.add(dgvaStudy.generateVariantStudy());
+            }
         }
         long end = System.currentTimeMillis();
         return new QueryResult(null, ((Long) (end - start)).intValue(), variantstudies.size(), variantstudies.size(), null, null, variantstudies);
     }
-
 
     @Override
     public QueryResult listStudies() {
@@ -62,12 +63,12 @@ public class StudyDgvaDBAdaptor implements StudyDBAdaptor {
     }
 
     @Override
-    public QueryResult getStudyById(String studyId, QueryOptions queryOptions) {
+    public QueryResult getStudyById(String s, QueryOptions queryOptions) {
         long start = System.currentTimeMillis();
-        DgvaStudyBrowser dgvaStudy = dgvaStudyBrowserRepository.getOne(studyId);
+        EvaStudyBrowser study = evaStudyBrowserRepository.findOne(s);
         List<VariantStudy> variantStudy = new ArrayList<>();
-        if (dgvaStudy != null) {
-            variantStudy.add(dgvaStudy.generateVariantStudy());
+        if (study != null) {
+            variantStudy.add(study.generateVariantStudy());
         }
         long end = System.currentTimeMillis();
         return new QueryResult(null, ((Long) (end - start)).intValue(), variantStudy.size(), variantStudy.size(), null, null, variantStudy);
