+++ eva-lib/src/main/java/uk/ac/ebi/eva/lib/metadata/eva/VariantSourceEvaProDBAdaptor.java
@@ -13,11 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.ac.ebi.eva.lib.metadata;
+package uk.ac.ebi.eva.lib.metadata.eva;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import uk.ac.ebi.eva.commons.core.models.stats.VariantSourceStats;
+import uk.ac.ebi.eva.lib.metadata.VariantSourceDBAdaptor;
 import uk.ac.ebi.eva.lib.models.FileFtpReference;
 import uk.ac.ebi.eva.lib.utils.QueryOptions;
 import uk.ac.ebi.eva.lib.utils.QueryResult;
