+++ eva-server/src/main/java/uk/ac/ebi/eva/server/ws/ga4gh/GA4GHVariantCallSetWSServer.java
@@ -35,8 +35,8 @@ import uk.ac.ebi.eva.lib.models.ga4gh.GACallSet;
 import uk.ac.ebi.eva.lib.models.ga4gh.GACallSetFactory;
 import uk.ac.ebi.eva.lib.models.ga4gh.GASearchCallSetsRequest;
 import uk.ac.ebi.eva.lib.models.ga4gh.GASearchCallSetsResponse;
-import uk.ac.ebi.eva.lib.utils.DBAdaptorConnector;
-import uk.ac.ebi.eva.lib.utils.MultiMongoDbFactory;
+import uk.ac.ebi.eva.lib.eva_utils.DBAdaptorConnector;
+import uk.ac.ebi.eva.lib.eva_utils.MultiMongoDbFactory;
 import uk.ac.ebi.eva.server.Utils;
 import uk.ac.ebi.eva.server.ws.EvaWSServer;
